+++ src/Message/types.ts
@@ -1,0 +1,5 @@
+export interface MessageProps {
+  message: string;
+  className?: string;
+  variant: "error" | "success" | "info";
+}