+++ src/RadioGroup/RadioGroup.tsx
@@ -1,0 +1,65 @@
+import { ChangeEvent, FocusEvent, useEffect, useState } from "react";
+import { ControlledRadio } from "../ControlledRadio";
+import { RadioGroupProps } from "./types";
+import styles from "./RadioGroup.module.scss";
+import Message from "../Message/Message";
+const RadioGroup = ({
+  options,
+  label,
+  disabled,
+  error,
+  onChange,
+  onBlur,
+  name,
+}: RadioGroupProps) => {
+  const { radioGroupContainer, radioGroupLabel, optionsContainer } = styles;
+  const [optionSelected, setOptionSelected] = useState<string>("");
+
+  useEffect(() => {
+    const optionWithInitialChecked = options.find(
+      ({ initialChecked }) => initialChecked
+    );
+    if (optionWithInitialChecked) {
+      setOptionSelected(optionWithInitialChecked.value);
+    }
+  }, [options]);
+
+  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
+    setOptionSelected(event.target.value);
+    onChange && onChange(event);
+  };
+
+  const handleBlur = (event: FocusEvent<HTMLInputElement, Element>) => {
+    onBlur && onBlur(event);
+  };
+
+  return (
+    <div className={radioGroupContainer}>
+      {/* LABEL */}
+      {label && <label className={radioGroupLabel}>{label}</label>}
+
+      {/* OPTONS */}
+      <div className={`${optionsContainer}`}>
+        {options.map(
+          ({ label, value, disabled: individualDisabled }, index) => (
+            <ControlledRadio
+              key={index}
+              checked={optionSelected === value}
+              label={label}
+              name={name ?? `controlledRadio${value}`}
+              value={value}
+              disabled={disabled || individualDisabled}
+              onChange={handleChange}
+              onBlur={handleBlur}
+            />
+          )
+        )}
+      </div>
+
+      {/* SHOW ERROR MESSAGE */}
+      {error && <Message message={error} variant="error" />}
+    </div>
+  );
+};
+
+export default RadioGroup;